backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
pymongo==4.13.0
django-cors-headers==4.7.0
python-dotenv==1.0.0
bcrypt==4.3.0 